+++ doc/whatsnew.rst
@@ -12,7 +12,11 @@ v0.6.dev
 
 - Update pyRiemann from Python 3.7 to 3.8. pr`254` by :user:`qbarthelemy`
 
-- Speedup pairwise distance function :func:`pyriemann.utils.distance.pairwise_distance` by adding individual functions for 'euclid', 'harmonic', 'logeuclid' and 'riemann' metrics. :pr:`256` by :user:`gabelstein`
+- Speedup pairwise distance function :func:`pyriemann.utils.distance.pairwise_distance`
+  by adding individual functions for 'euclid', 'harmonic', 'logeuclid' and 'riemann' metrics. :pr:`256` by :user:`gabelstein`
+
+- Add :func:`pyriemann.utils.test.is_real_type` to check the type of input arrays and
+  add :func:`pyriemann.utils.covariance.covariance_scm` allowing to process complex-valued inputs for 'scm' covariance estimator. :pr:`251` by :user:`qbarthelemy`
 
 
 v0.5 (Jun 2023)